Refactor CAEP and RISC classes into subpackges
to keep the top package directory from being so messy.
