+++ src/main/java/com/sailpoint/sse/model/caep/CAEPSessionRevoked.java
@@ -4,7 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.sailpoint.sse.model;
+package com.sailpoint.sse.model.caep;
+
+import com.sailpoint.sse.model.SSEventTypes;
 
 public class CAEPSessionRevoked extends CAEPBaseEvent {
