+++ src/main/java/com/sailpoint/sse/model/caep/CAEPTokenClaimsChange.java
@@ -4,9 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.sailpoint.sse.model;
+package com.sailpoint.sse.model.caep;
 
 import com.nimbusds.jose.shaded.json.JSONObject;
+import com.sailpoint.sse.model.SSEventTypes;
+import com.sailpoint.sse.model.ValidationException;
 
 public class CAEPTokenClaimsChange extends CAEPBaseEvent {
