+++ src/main/java/com/sailpoint/sse/model/caep/NISTAuthenticatorAssuranceLevel.java
@@ -4,32 +4,32 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.sailpoint.sse.model;
+package com.sailpoint.sse.model.caep;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public enum CAEPInitiatingEntity {
-    ADMIN("admin"),
-    USER("user"),
-    POLICY("policy"),
-    SYSTEM("system");
+public enum NISTAuthenticatorAssuranceLevel {
 
-    private static final Map<String, CAEPInitiatingEntity> BY_NAME = new HashMap<>();
+    NIST_AAL1("nist-aal1"),
+    NIST_AAL2("nist-aal2"),
+    NIST_AAL3("nist-aal3");
+
+    private static final Map<String, NISTAuthenticatorAssuranceLevel> BY_NAME = new HashMap<>();
 
     static {
-        for (CAEPInitiatingEntity t : values()) {
+        for (NISTAuthenticatorAssuranceLevel t : values()) {
             BY_NAME.put(t.name, t);
         }
     }
 
     private final String name;
 
-    CAEPInitiatingEntity(String s) {
+    NISTAuthenticatorAssuranceLevel(final String s) {
         name = s;
     }
 
-    public static CAEPInitiatingEntity valueOfLabel(String name) {
+    public static NISTAuthenticatorAssuranceLevel valueOfLabel(String name) {
         return BY_NAME.get(name);
     }
 
@@ -37,7 +37,7 @@ public enum CAEPInitiatingEntity {
         return BY_NAME.containsKey(name);
     }
 
-    public boolean equalsName(String otherName) {
+    public boolean equalsName(final String otherName) {
         return name.equals(otherName);
     }
