+++ src/main/java/com/sailpoint/sse/model/risc/RISCAccountCredentialChangeRequired.java
@@ -4,7 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.sailpoint.sse.model;
+package com.sailpoint.sse.model.risc;
+
+import com.sailpoint.sse.model.SSEvent;
+import com.sailpoint.sse.model.SSEventTypes;
 
 public class RISCAccountCredentialChangeRequired extends SSEvent {
