+++ src/main/java/com/sailpoint/sse/model/risc/RISCIdentifierChanged.java
@@ -4,7 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.sailpoint.sse.model;
+package com.sailpoint.sse.model.risc;
+
+import com.sailpoint.sse.model.SSEvent;
+import com.sailpoint.sse.model.SSEventTypes;
+import com.sailpoint.sse.model.ValidationException;
 
 public class RISCIdentifierChanged extends SSEvent {
