+++ src/test/java/com/sailpoint/sse/model/CAEPAssuranceLevelChangeTest.java
@@ -10,6 +10,10 @@ import com.nimbusds.jose.shaded.json.JSONObject;
 import com.nimbusds.jose.util.JSONObjectUtils;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.util.DateUtils;
+import com.sailpoint.sse.model.caep.CAEPAssuranceLevelChange;
+import com.sailpoint.sse.model.caep.CAEPAssuranceLevelChangeDirection;
+import com.sailpoint.sse.model.caep.CAEPInitiatingEntity;
+import com.sailpoint.sse.model.caep.NISTAuthenticatorAssuranceLevel;
 import org.junit.Test;
 
 import java.text.ParseException;
@@ -32,7 +36,7 @@ public class CAEPAssuranceLevelChangeTest {
         CAEPAssuranceLevelChange evt = new CAEPAssuranceLevelChange.Builder()
                 .currentLevel(NISTAuthenticatorAssuranceLevel.NIST_AAL2)
                 .previousLevel(NISTAuthenticatorAssuranceLevel.NIST_AAL1)
-                .changeDirection(AssuranceLevelChangeDirection.INCREASE)
+                .changeDirection(CAEPAssuranceLevelChangeDirection.INCREASE)
                 .eventTimestamp(1615304991643L)
                 .initiatingEntity(CAEPInitiatingEntity.USER)
                 .subject(subj)
