+++ src/test/java/com/sailpoint/sse/model/CAEPCredentialChangeTest.java
@@ -10,6 +10,10 @@ import com.nimbusds.jose.shaded.json.JSONObject;
 import com.nimbusds.jose.util.JSONObjectUtils;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.util.DateUtils;
+import com.sailpoint.sse.model.caep.CAEPChangeType;
+import com.sailpoint.sse.model.caep.CAEPCredentialChange;
+import com.sailpoint.sse.model.caep.CAEPCredentialType;
+import com.sailpoint.sse.model.caep.CAEPInitiatingEntity;
 import org.junit.Test;
 
 import java.text.ParseException;
