+++ src/test/java/com/sailpoint/sse/model/CAEPDeviceComplianceChangeTest.java
@@ -10,6 +10,9 @@ import com.nimbusds.jose.shaded.json.JSONObject;
 import com.nimbusds.jose.util.JSONObjectUtils;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.util.DateUtils;
+import com.sailpoint.sse.model.caep.CAEPComplianceStatus;
+import com.sailpoint.sse.model.caep.CAEPDeviceComplianceChange;
+import com.sailpoint.sse.model.caep.CAEPInitiatingEntity;
 import org.junit.Test;
 
 import java.text.ParseException;
