+++ src/test/java/com/sailpoint/sse/model/CAEPSessionRevokedTest.java
@@ -10,6 +10,8 @@ import com.nimbusds.jose.shaded.json.JSONObject;
 import com.nimbusds.jose.util.JSONObjectUtils;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.util.DateUtils;
+import com.sailpoint.sse.model.caep.CAEPInitiatingEntity;
+import com.sailpoint.sse.model.caep.CAEPSessionRevoked;
 import org.junit.Test;
 
 import java.text.ParseException;
