+++ src/test/java/com/sailpoint/sse/model/OpenIDSSEProfileTest.java
@@ -10,6 +10,9 @@ import com.nimbusds.jose.shaded.json.JSONObject;
 import com.nimbusds.jose.util.JSONObjectUtils;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.util.DateUtils;
+import com.sailpoint.sse.model.caep.CAEPSessionRevoked;
+import com.sailpoint.sse.model.caep.CAEPTokenClaimsChange;
+import com.sailpoint.sse.model.risc.RISCAccountEnabled;
 import org.junit.Test;
 
 import java.text.ParseException;
